% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_functions.R
\name{plot.comp.funcs.res.dt}
\alias{plot.comp.funcs.res.dt}
\title{plot.comp.funcs.res.dt}
\usage{
\method{plot}{comp.funcs.res.dt}(comp_funcs_res, y)
}
\arguments{
\item{comp_funcs_res}{Object of class \code{comp.funcs.res.dt} - result of function \code{compare_functions()}}

\item{y}{\code{character} value of column to use for Y axis}
}
\value{
ggplot2 object with the plot
}
\description{
Gives a plot object for specified Y axis
}
\details{
S3 \code{plot} method for class \code{comp.funcs.res.dt}. Plots a line chart with function_index split by function_index with \code{size} in X axis
}
\author{
Sujeet G Pillai
}
\keyword{compare}
\keyword{functions,}
\keyword{ggplot}
\keyword{method,}
\keyword{plot,}
\keyword{s3}
