% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compare_functions_app.R
\name{run_compare_functions_app}
\alias{run_compare_functions_app}
\title{compare_functions}
\usage{
run_compare_functions_app()
}
\description{
Launches a shiny app that runs on 'compareFunctions'
}
\details{
Users can load custom parameters and compare between 2-6 functions at the same time to get summary results, plots and regression outputs of the comparison.
}
\author{
Sujeet G Pillai
}
\keyword{app,}
\keyword{compare}
\keyword{example}
\keyword{functions,}
