% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_functions.R
\name{summary.comp.funcs.res.dt}
\alias{summary.comp.funcs.res.dt}
\title{summary.comp.funcs.res.dt}
\usage{
\method{summary}{comp.funcs.res.dt}(comp_funcs_res)
}
\arguments{
\item{comp_funcs_res}{Object of class \code{comp.funcs.res.dt} - result of function \code{compare_functions()}}
}
\description{
Gives a summary by function index
}
\details{
S3 \code{summary} method for class \code{comp.funcs.res.dt}. Prints traditional summary output by function index
}
\author{
Sujeet G Pillai
}
\keyword{compare}
\keyword{functions,}
\keyword{method}
\keyword{s3}
\keyword{summary,}
